+++ client/src/components/reviews/Reviews.js
@@ -4,7 +4,7 @@ import { getReviews } from "../../actions";
 import ReviewForm from "./ReviewForm";
 import Review from "./Review";
 import ReviewBars from './ReviewBars';
-import "../../reviews.scss";
+import "./reviews.scss";
 
 class Reviews extends React.Component {
   state = {
